feat: header layout style / components 분리 #13
