+++ src/layouts/Header/HeaderNav/HeaderNav.style.tsx
@@ -1,37 +1,6 @@
-import { useLogout } from "@/utils/hooks";
+import { HeaderVisibleProps } from "@/types/headerStyle";
 import { NavLink } from "react-router-dom";
 import styled, { css, keyframes } from "styled-components";
-import { HiOutlineHome } from "@react-icons/all-files/hi/HiOutlineHome";
-import { FaRegUserCircle } from "@react-icons/all-files/fa/FaRegUserCircle";
-import { RiLogoutBoxRLine } from "@react-icons/all-files/ri/RiLogoutBoxRLine";
-import { BiSearchAlt2 } from "@react-icons/all-files/bi/BiSearchAlt2";
-import { HeaderVisibleProps } from "@/types/headerStyle";
-
-const HeaderNav = ({ onToggleSearch, visible, render }) => {
-  const { logout } = useLogout();
-
-  const onLogout = async () => {
-    await logout();
-  };
-
-  return (
-    <StyledHeaderNav>
-      <StyledSearch visible={visible} render={render}>
-        <BiSearchAlt2 onClick={onToggleSearch} />
-      </StyledSearch>
-      <StyledHome to="/" visible={visible} render={render}>
-        <HiOutlineHome />
-      </StyledHome>
-      <StyledProfile to="/profile" visible={visible} render={render}>
-        <FaRegUserCircle />
-      </StyledProfile>
-      <StyledLogOut visible={visible} render={render}>
-        <RiLogoutBoxRLine onClick={() => onLogout()} />
-      </StyledLogOut>
-    </StyledHeaderNav>
-  );
-};
-
 
 const slideIn = keyframes`
   from{
@@ -134,4 +103,4 @@ const StyledLogOut = styled(StyledNavItem)`
   `}
 `;
 
-export default HeaderNav;
+export { StyledHeaderNav, StyledHome, StyledLogOut, StyledProfile, StyledSearch };
