+++ src/layouts/Header/SearchList/SearchList.tsx
@@ -1,0 +1,11 @@
+import { StyledSearchList } from "./SearchList.style";
+
+const SearchList = ({ search }) => {
+  return (
+    <StyledSearchList>
+      {search.result}
+    </StyledSearchList>
+  );
+};
+
+export default SearchList;
